fix: pytest failed in github ACTION (igsparser module not found)
When move to pyproject.toml, also clean up the requirements.txt.
There is "./" in the requirements.txt in previous version, that
cause the test env did install this package before running the pytest.

One posible solution is to install this package before pytest. But
I choice to add tests/context.py to add search path for the tests.
Thus I can run the pytest without installation.

Alos, add more info in pyproject.toml for this package.

Signed-off-by: Tim Chao <tim.chao@ingics.com>